+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for high-content screening (HCS) workflows with NGFF 0.5 metadata (#153)
 - New API function for retrieving the distinct array keys from a `ZarrStreamSettings` object (#154)
 
+### Fixed
+
+- A bug affecting the Zarr V3 writer that caused it to skip writing the chunk table on the final shard file when the
+  shard was not completely full on shutdown (#159)
+
 ## [0.5.2] - [2025-08-07](https://github.com/acquire-project/acquire-zarr/compare/v0.5.1...v0.5.2)
 
 ### Added